+++ gb_core/gameboy.h
@@ -16,7 +16,7 @@ class gameboy {
 
     void load_rom(uint8_t *romData, uint32_t romLength, uint8_t *ram, uint32_t ramLength);
 
-    void save_state(std::function<uint8_t *(uint32_t)> functor);
+    void save_state(std::function<uint8_t *(size_t)> functor);
     void load_state(uint8_t *state);
     void save_sram(std::function<void(uint8_t *, uint32_t)> functor);
